Added type hints to `@rules_python//python/runfiles`

+++ tests/runfiles/runfiles_test.py
@@ -1,4 +1,3 @@
-# pylint: disable=g-bad-file-header
 # Copyright 2018 The Bazel Authors. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,18 +15,20 @@
 import os
 import tempfile
 import unittest
+from typing import Any, List, Optional
 
 from python.runfiles import runfiles
 
 
 class RunfilesTest(unittest.TestCase):
-    # """Unit tests for `runfiles.Runfiles`."""
+    """Unit tests for `rules_python.python.runfiles.Runfiles`."""
 
-    def testRlocationArgumentValidation(self):
+    def testRlocationArgumentValidation(self) -> None:
         r = runfiles.Create({"RUNFILES_DIR": "whatever"})
-        self.assertRaises(ValueError, lambda: r.Rlocation(None))
+        assert r is not None
+        self.assertRaises(ValueError, lambda: r.Rlocation(None))  # type: ignore
         self.assertRaises(ValueError, lambda: r.Rlocation(""))
-        self.assertRaises(TypeError, lambda: r.Rlocation(1))
+        self.assertRaises(TypeError, lambda: r.Rlocation(1))  # type: ignore
         self.assertRaisesRegex(
             ValueError, "is not normalized", lambda: r.Rlocation("../foo")
         )
@@ -61,7 +62,7 @@ class RunfilesTest(unittest.TestCase):
             lambda: r.Rlocation("\\foo"),
         )
 
-    def testCreatesManifestBasedRunfiles(self):
+    def testCreatesManifestBasedRunfiles(self) -> None:
         with _MockFile(contents=["a/b c/d"]) as mf:
             r = runfiles.Create(
                 {
@@ -70,10 +71,11 @@ class RunfilesTest(unittest.TestCase):
                     "TEST_SRCDIR": "always ignored",
                 }
             )
+            assert r is not None
             self.assertEqual(r.Rlocation("a/b"), "c/d")
             self.assertIsNone(r.Rlocation("foo"))
 
-    def testManifestBasedRunfilesEnvVars(self):
+    def testManifestBasedRunfilesEnvVars(self) -> None:
         with _MockFile(name="MANIFEST") as mf:
             r = runfiles.Create(
                 {
@@ -81,6 +83,7 @@ class RunfilesTest(unittest.TestCase):
                     "TEST_SRCDIR": "always ignored",
                 }
             )
+            assert r is not None
             self.assertDictEqual(
                 r.EnvVars(),
                 {
@@ -97,6 +100,7 @@ class RunfilesTest(unittest.TestCase):
                     "TEST_SRCDIR": "always ignored",
                 }
             )
+            assert r is not None
             self.assertDictEqual(
                 r.EnvVars(),
                 {
@@ -117,6 +121,7 @@ class RunfilesTest(unittest.TestCase):
                     "TEST_SRCDIR": "always ignored",
                 }
             )
+            assert r is not None
             self.assertDictEqual(
                 r.EnvVars(),
                 {
@@ -126,23 +131,25 @@ class RunfilesTest(unittest.TestCase):
                 },
             )
 
-    def testCreatesDirectoryBasedRunfiles(self):
+    def testCreatesDirectoryBasedRunfiles(self) -> None:
         r = runfiles.Create(
             {
                 "RUNFILES_DIR": "runfiles/dir",
                 "TEST_SRCDIR": "always ignored",
             }
         )
+        assert r is not None
         self.assertEqual(r.Rlocation("a/b"), "runfiles/dir/a/b")
         self.assertEqual(r.Rlocation("foo"), "runfiles/dir/foo")
 
-    def testDirectoryBasedRunfilesEnvVars(self):
+    def testDirectoryBasedRunfilesEnvVars(self) -> None:
         r = runfiles.Create(
             {
                 "RUNFILES_DIR": "runfiles/dir",
                 "TEST_SRCDIR": "always ignored",
             }
         )
+        assert r is not None
         self.assertDictEqual(
             r.EnvVars(),
             {
@@ -151,13 +158,13 @@ class RunfilesTest(unittest.TestCase):
             },
         )
 
-    def testFailsToCreateManifestBasedBecauseManifestDoesNotExist(self):
+    def testFailsToCreateManifestBasedBecauseManifestDoesNotExist(self) -> None:
         def _Run():
             runfiles.Create({"RUNFILES_MANIFEST_FILE": "non-existing path"})
 
         self.assertRaisesRegex(IOError, "non-existing path", _Run)
 
-    def testFailsToCreateAnyRunfilesBecauseEnvvarsAreNotDefined(self):
+    def testFailsToCreateAnyRunfilesBecauseEnvvarsAreNotDefined(self) -> None:
         with _MockFile(contents=["a b"]) as mf:
             runfiles.Create(
                 {
@@ -175,7 +182,7 @@ class RunfilesTest(unittest.TestCase):
         self.assertIsNone(runfiles.Create({"TEST_SRCDIR": "always ignored"}))
         self.assertIsNone(runfiles.Create({"FOO": "bar"}))
 
-    def testManifestBasedRlocation(self):
+    def testManifestBasedRlocation(self) -> None:
         with _MockFile(
             contents=[
                 "Foo/runfile1",
@@ -205,7 +212,7 @@ class RunfilesTest(unittest.TestCase):
             else:
                 self.assertEqual(r.Rlocation("/foo"), "/foo")
 
-    def testManifestBasedRlocationWithRepoMappingFromMain(self):
+    def testManifestBasedRlocationWithRepoMappingFromMain(self) -> None:
         with _MockFile(
             contents=[
                 ",config.json,config.json~1.2.3",
@@ -280,7 +287,7 @@ class RunfilesTest(unittest.TestCase):
             self.assertIsNone(r.Rlocation("my_module", ""))
             self.assertIsNone(r.Rlocation("protobuf", ""))
 
-    def testManifestBasedRlocationWithRepoMappingFromOtherRepo(self):
+    def testManifestBasedRlocationWithRepoMappingFromOtherRepo(self) -> None:
         with _MockFile(
             contents=[
                 ",config.json,config.json~1.2.3",
@@ -362,7 +369,7 @@ class RunfilesTest(unittest.TestCase):
             self.assertIsNone(r.Rlocation("my_module", "protobuf~3.19.2"))
             self.assertIsNone(r.Rlocation("protobuf", "protobuf~3.19.2"))
 
-    def testDirectoryBasedRlocation(self):
+    def testDirectoryBasedRlocation(self) -> None:
         # The _DirectoryBased strategy simply joins the runfiles directory and the
         # runfile's path on a "/". This strategy does not perform any normalization,
         # nor does it check that the path exists.
@@ -374,7 +381,7 @@ class RunfilesTest(unittest.TestCase):
         else:
             self.assertEqual(r.Rlocation("/foo"), "/foo")
 
-    def testDirectoryBasedRlocationWithRepoMappingFromMain(self):
+    def testDirectoryBasedRlocationWithRepoMappingFromMain(self) -> None:
         with _MockFile(
             name="_repo_mapping",
             contents=[
@@ -441,7 +448,7 @@ class RunfilesTest(unittest.TestCase):
 
             self.assertEqual(r.Rlocation("config.json", ""), dir + "/config.json")
 
-    def testDirectoryBasedRlocationWithRepoMappingFromOtherRepo(self):
+    def testDirectoryBasedRlocationWithRepoMappingFromOtherRepo(self) -> None:
         with _MockFile(
             name="_repo_mapping",
             contents=[
@@ -513,40 +520,49 @@ class RunfilesTest(unittest.TestCase):
                 r.Rlocation("config.json", "protobuf~3.19.2"), dir + "/config.json"
             )
 
-    def testCurrentRepository(self):
+    def testCurrentRepository(self) -> None:
         # Under bzlmod, the current repository name is the empty string instead
         # of the name in the workspace file.
         if bool(int(os.environ["BZLMOD_ENABLED"])):
             expected = ""
         else:
             expected = "rules_python"
-        self.assertEqual(
-            runfiles.Create({"RUNFILES_DIR": "whatever"}).CurrentRepository(), expected
-        )
+        r = runfiles.Create({"RUNFILES_DIR": "whatever"})
+        assert r is not None
+        self.assertEqual(r.CurrentRepository(), expected)
 
     @staticmethod
-    def IsWindows():
+    def IsWindows() -> bool:
         return os.name == "nt"
 
 
-class _MockFile(object):
-    def __init__(self, name=None, contents=None):
+class _MockFile:
+    def __init__(
+        self, name: Optional[str] = None, contents: Optional[List[Any]] = None
+    ) -> None:
         self._contents = contents or []
         self._name = name or "x"
-        self._path = None
+        self._path: Optional[str] = None
 
-    def __enter__(self):
+    def __enter__(self) -> Any:
         tmpdir = os.environ.get("TEST_TMPDIR")
         self._path = os.path.join(tempfile.mkdtemp(dir=tmpdir), self._name)
         with open(self._path, "wt") as f:
             f.writelines(l + "\n" for l in self._contents)
         return self
 
-    def __exit__(self, exc_type, exc_value, traceback):
-        os.remove(self._path)
-        os.rmdir(os.path.dirname(self._path))
+    def __exit__(
+        self,
+        exc_type: Any,  # pylint: disable=unused-argument
+        exc_value: Any,  # pylint: disable=unused-argument
+        traceback: Any,  # pylint: disable=unused-argument
+    ) -> None:
+        if self._path:
+            os.remove(self._path)
+            os.rmdir(os.path.dirname(self._path))
 
-    def Path(self):
+    def Path(self) -> str:
+        assert self._path is not None
         return self._path
